% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{aggreg}
\alias{aggreg}
\title{Aggregate xts object}
\usage{
aggreg(daxts, FUN, freq = "daily", dig = 4, ...)
}
\arguments{
\item{daxts}{object of calss \code{xts} with simple or multiple columns}

\item{FUN}{a function defined by the user.}

\item{freq}{a char string to the aggregation. It can be daily, weekly, monthly, quarterly or yearly}

\item{dig}{number of output digits}
}
\value{
a object from class \code{xts} with the data aggregated
}
\description{
Aggregate xts object by a given function. It can be any function defined by the user. Actually it's possible to aggregate data by day, month, week, quarter and year.
}
\examples{
data(diario)
d1 <- cvforecast::ConvertData(diario,  tsfrequency = "day", OutType = "xts", OutlierClean = FALSE)
fun1 <- function(x) {c(mean(x, na.rm=TRUE) + sd(x,na.rm=TRUE))}
aggreg(d1[,1:5], fun1, "monthly")
aggreg(d1[,1:5], mean, "monthly")
}
\seealso{
\code{\link{apply.daily}}
}

