% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{auto.arimaForecast}
\alias{auto.arimaForecast}
\title{auto.arima forecast wrapper}
\usage{
auto.arimaForecast(x, h, level = 95, onlyfc = TRUE, xreg = NULL,
  newxreg = NULL, ...)
}
\description{
auto.arima forecast wrapper
}

