% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{ConvertDataToTs}
\alias{ConvertDataToTs}
\title{Conver data to ts or xts}
\usage{
ConvertDataToTs(Data, tsfrequency = "month", OutType = "ts",
  OutlierClean = TRUE, ...)
}
\description{
Conver data to ts or xts
}

