% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{cvBestModel}
\alias{cvBestModel}
\title{Best forecast among several ones}
\usage{
cvBestModel(objcvfore, cvMethod = "MAPE", residlevel = 0.1, ...)
}
\arguments{
\item{objcvfore}{'cvforecst' object}

\item{cvMethod}{gof statistic, MAPE, MASE, MAE, etc. \code{\link{accuracy}}}

\item{residlevel}{confidence level for residuals tests}
}
\description{
This function receives a list of forecast models from the class 'cvforecst' and returns a list of their characteristics.
}
\examples{
#Define cross validation parameters
myControl <- cvForecastControl(
minObs = 14,
stepSize = 1,
maxHorizon = 30,
residlevel = 0.05)

#Paralell execution improves the processing time
#cl <- makeCluster(4, type='SOCK')
#registerDoParallel(cl)

#Load data
x <- AirPassengers
fit <- cvforecast(x, myControl)
cvBestModel(fit)
}
\author{
LOPES, J. E.
}

