% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{cvForecastControl}
\alias{cvForecastControl}
\title{Default Cross-validation control}
\usage{
cvForecastControl(stepSize = 1, maxHorizon = 1, minObs = 2 *
  max(cycle(x)), fixedWindow = TRUE, summaryFunc = tsSummary,
  preProcess = FALSE, ppMethod = "guerrero", cvMethod = "MAPE",
  tsfrequency = "month", OutlierClean = TRUE)
}
\description{
Default Cross-validation control
}

