% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{Desc}
\alias{Desc}
\title{Descriptive statistics}
\usage{
Desc(dados, nivel = 0.95, tipoci = "basic", nsimu = 500, dig = 2)
}
\arguments{
\item{dados}{data.frame, matrix or numeric vector}

\item{nivel}{significance level for bootstrap non-paramentric ci for the mean}

\item{tipoci}{type of confidence interval test. See \code{\link{boot.ci}}}

\item{nsimu}{bootstrap simulations number}

\item{dig}{number of digits for output}
}
\value{
data.frame with min, mean, median, max, sd, sd + mean, coefficeint of variation and bootstrap non-parametric confidence interval for the mean
}
\description{
Receives a data.frame, vector or matrix and computes descriptive statistics.
}
\examples{
Desc(cars)
Desc(rnorm(100, 2, 3))
}

