% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{descritiva}
\alias{descritiva}
\title{Descriptive analysis}
\usage{
descritiva(x, basic = TRUE, desc = TRUE, norm = FALSE, p = 0.95,
  dig = 6)
}
\description{
Descriptive analysis
}

