% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{DetectSeasonality}
\alias{DetectSeasonality}
\title{Seasolaity detection test}
\usage{
DetectSeasonality(x, level = 0.95)
}
\arguments{
\item{x}{'ts' data to be tested}

\item{level}{significance value to test the null hypothesis that the data has a seasonal pattern.}
}
\value{
bolean TRUE or FALSE
}
\description{
This function perform a simple test for seasonality test based on http://robjhyndman.com/hyndsight/detecting-seasonality/.
The idea is to perform a chi-squared test for the deviance on two forecst models of type ETS. The first model is an automated ETS model and the second is a ETS model with fixed seasonal component. The NULL is that the data has a seasonal component. If chisq p.value is greater than (1-level) then reject NULL.
}
\note{
Due this function performs two ETS forecast inside it. It can be slow.
}
\examples{
par(mfrow = c(2,1))
plot(AirPassengers)
plot(lynx)
DetectSeasonality(AirPassengers)
DetectSeasonality(lynx)
par(mfrow = c(1,1))
}
\author{
LOPES, J. E.
}

