% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{fc_hwes}
\alias{fc_hwes}
\title{HoltWinters Exponential Smoothing forecast wrapper}
\usage{
fc_hwes(x, h, level = 95, onlyfc = TRUE, ...)
}
\arguments{
\item{x}{'ts' data}

\item{h}{forecast horizon}

\item{level}{confidence level. Default is 0.95}

\item{onlyfc}{if TRUE return only forecasts, otherwise returns a full forecast classed object}

\item{...}{extra args, if needed.}
}
\value{
forecasts from ts data or an object of class 'forecast'
}
\description{
HoltWinters Exponential Smoothing forecast wrapper
}
\examples{
fit <- fc_hwes(AirPassengers, h=10, onlyfc = FALSE)
plot(fit)
Mresid(fit)
tsSummary(fit)
}

