% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{fc_mean}
\alias{fc_mean}
\title{Mean forecast wrapper}
\usage{
fc_mean(x, h, level = 95, onlyfc = TRUE, ...)

fc_mean(x, h, level = 95, onlyfc = TRUE, ...)
}
\arguments{
\item{x}{'ts' data}

\item{h}{forecast horizon}

\item{level}{confidence level. Default is 0.95}

\item{onlyfc}{if TRUE return only forecasts, otherwise returns a full forecast classed object}

\item{...}{extra args, if needed.}

\item{x}{'ts' data}

\item{h}{forecast horizon}

\item{level}{confidence level. Default is 0.95}

\item{onlyfc}{if TRUE return only forecasts, otherwise returns a full forecast classed object}

\item{...}{extra args, if needed.}
}
\value{
forecasts from ts data or an object of class 'forecast'

forecasts from ts data or an object of class 'forecast'
}
\description{
Mean forecast wrapper

Meanf forecast wrapper
}
\examples{
fit <- fc_mean(AirPassengers, h=10, onlyfc = FALSE)
plot(fit)
Mresid(fit)
tsSummary(fit)
fit <- fc_mean(AirPassengers, h=10, onlyfc = FALSE)
plot(fit)
Mresid(fit)
tsSummary(fit)
}

