% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{fc_nnetar}
\alias{fc_nnetar}
\title{NNetar forecast wrapper}
\usage{
fc_nnetar(x, h, level = 95, onlyfc = TRUE, nn_p = 1, ...)
}
\arguments{
\item{x}{'ts' data}

\item{h}{forecast horizon}

\item{level}{confidence level. Default is 0.95}

\item{onlyfc}{if TRUE return only forecasts, otherwise returns a full forecast classed object}

\item{nn_p}{Embedding dimension for non-seasonal time series. Number of non-seasonal lags used as inputs. For non-seasonal time series, the default is the optimal number of lags (according to the AIC) for a linear AR(p) model. For seasonal time series, the same method is used but applied to seasonally adjusted data (from an stl decomposition) - From Rob J Hyndman.}

\item{...}{extra args, if needed.}
}
\value{
forecasts from ts data or an object of class 'forecast'
}
\description{
NNetar forecast wrapper
}
\examples{
fit <- fc_nnetar(AirPassengers, h=10, onlyfc = FALSE)
plot(fit)
Mresid(fit)
tsSummary(fit)
}

