% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{fc_stl}
\alias{fc_stl}
\title{Stl forecast wrapper}
\usage{
fc_stl(x, h, level = 95, method = "arima", onlyfc = TRUE, ...)
}
\arguments{
\item{x}{'ts' data}

\item{h}{forecast horizon}

\item{level}{confidence level. Default is 0.95}

\item{method}{Method to use for forecasting the seasonally adjusted series.}

\item{onlyfc}{if TRUE return only forecasts, otherwise returns a full forecast classed object}

\item{...}{extra args, if needed.}
}
\value{
forecasts from ts data or an object of class 'forecast'
}
\description{
Stl forecast wrapper
}
\examples{
fit <- fc_stl(AirPassengers, h=10, onlyfc = FALSE)
plot(fit)
Mresid(fit)
tsSummary(fit)
}

