% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{fc_theta}
\alias{fc_theta}
\title{Theta forecast wrapper}
\usage{
fc_theta(x, h, level = 95, onlyfc = TRUE, ...)
}
\arguments{
\item{x}{'ts' data}

\item{h}{forecast horizon}

\item{level}{confidence level. Default is 0.95}

\item{onlyfc}{if TRUE return only forecasts, otherwise returns a full forecast classed object}

\item{...}{extra args, if needed.}
}
\value{
forecasts from ts data or an object of class 'forecast'
}
\description{
Theta forecast wrapper
}
\examples{
fit <- fc_theta(AirPassengers, h=10, onlyfc = FALSE)
plot(fit)
Mresid(fit)
tsSummary(fit)
}

