% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{ForecastHorizon}
\alias{ForecastHorizon}
\title{Generate forecast horizon time sequency}
\usage{
ForecastHorizon(XtsData, tsfrequency, horizon)
}
\description{
Generate forecast horizon time sequency
}

