% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{forecastMethod}
\alias{forecastMethod}
\title{Find best forecast methods based on trend and linearity}
\usage{
forecastMethod(x)
}
\arguments{
\item{x}{ts object, univariate time series}
}
\description{
Automatically estimates the most adequate forecasts method for a 'ts' object based on linearity and trend tests and return a list containing all functions names. See \code{\link{nparTrend}} and \code{\link{linearityTest}} for more.
}
\examples{
forecastMethod(lynx)
forecastMethod(AirPassengers)
}
\author{
LOPES, J. E.
}

