% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{HWesForecast}
\alias{HWesForecast}
\title{HoltWinters Exponential Smoothing forecast wrapper}
\usage{
HWesForecast(x, h, level = 95, onlyfc = TRUE, ...)
}
\description{
HoltWinters Exponential Smoothing forecast wrapper
}

