% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{HWnsForecast}
\alias{HWnsForecast}
\title{HoltWinters Non Seazonal forecast wrapper}
\usage{
HWnsForecast(x, h, level = 95, onlyfc = TRUE, ...)
}
\description{
HoltWinters Non Seazonal forecast wrapper
}

