% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{HWsForecast}
\alias{HWsForecast}
\title{HoltWinters Sazonal forecast wrapper by stats}
\usage{
HWsForecast(x, h, level = 95, onlyfc = TRUE, ...)
}
\description{
HoltWinters Sazonal forecast wrapper by stats
}

