% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{linearityTest}
\alias{linearityTest}
\title{Linearity tests}
\usage{
linearityTest(x, Test, all = FALSE)
}
\arguments{
\item{x}{'ts' data}

\item{Test}{string containing test's names. They are "terasvirta", "white", "keenan", "tsay", "tarTest".}
}
\value{
data.frame with tests statistics and p.values
}
\description{
Compute six linearity tests among Terasvirta, White, Keenan, McleodLi, Tsay and Tar.
}
\examples{
linearityTest(AirPassengers)
linearityTest(AirPassengers, Test="tsay")
}
\author{
LOPES, J. E.
}
\seealso{
See \code{\link{terasvirta.test}}, \code{\link{white.test}},\code{\link{Keenan.test}},\code{\link{Tsay.test}} and \code{\link{tlrt}}
}

