% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{LjungBtest_Acuracia}
\alias{LjungBtest_Acuracia}
\title{Box and Cox tests and Ljung}
\usage{
LjungBtest_Acuracia(model, ...)
}
\arguments{
\item{model}{A forecast model. See \code{\link{cvts2}}.}

\item{...}{extra args, if needed.}
}
\description{
Performs Ljung and Box test and also Durbin-Watson both for autocorrelation on the residuals from the model. It returns also accuracy statistics. See \code{\link{accuracy}}, \code{\link{Box.test}}, \code{\link{dwtest}}.
}
\examples{
x <- AirPassengers
fit <- fc_auto.arima(x, h = 10, onlyfc=FALSE)
LjungBtest_Acuracia(fit)
plot(fit)
}
\author{
LOPES, J. E.
}

