% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{lm_press_stat}
\alias{lm_press_stat}
\title{Compute R^2 and PRESS statistic form lm}
\usage{
lm_press_stat(obj)
}
\arguments{
\item{obj}{object from \code{\link{lm}} class}
}
\value{
data.frama with test statistics
}
\description{
Compute R^2 and PRESS statistic form lm
}
\examples{
fit <- step(lm(mpg~., data=mtcars), trace = 0)
lm_press_stat(fit)
}

