% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{lmForecast}
\alias{lmForecast}
\title{Linear model forecast wrapper}
\usage{
lmForecast(x, h, level = 95, onlyfc = TRUE, xreg = NULL, newxreg = NULL,
  ...)
}
\description{
Linear model forecast wrapper
}

