% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{meanForecast}
\alias{meanForecast}
\title{Mean forecast wrapper}
\usage{
meanForecast(x, h, level = 95, onlyfc = TRUE, ...)

meanForecast(x, h, level = 95, onlyfc = TRUE, ...)
}
\description{
Mean forecast wrapper

Meanf forecast wrapper
}

