% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{nnetarForecast}
\alias{nnetarForecast}
\title{NNetar forecast wrapper}
\usage{
nnetarForecast(x, h, level = 95, onlyfc = TRUE, nn_p = 1, ...)
}
\description{
NNetar forecast wrapper
}

