% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{ordem_diferencas}
\alias{ordem_diferencas}
\title{Autofind time sieries number of differences}
\usage{
ordem_diferencas(x, alpha = 0.05, plot = TRUE, ...)
}
\description{
Autofind time sieries number of differences
}

