% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{plot.cvforecast}
\alias{plot.cvforecast}
\title{Default plot for Cross-validation forecast}
\usage{
plot.cvforecast(obj, cv_forecast = TRUE, ...)
}
\description{
Default plot for Cross-validation forecast
}

