% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{plotcv}
\alias{plotcv}
\title{Plot forecasts from cvforecast objetc}
\usage{
plotcv(obj, x_labgraf = "Di<e1>ria", v_titgraf = "Hist<f3>rico mais Proje<e7><e3>o",
  ...)
}
\arguments{
\item{obj}{object of cvforecast class}

\item{x_labgraf}{label for x-axis}

\item{v_titgraf}{plot title#'}
}
\description{
This function makes plots forecasts from cvforecast objects by using \code{\link{ggplot2}} framework. This function different from \code{plot.cvforecast} try to plot historical and forecast data including dates in the x axis.
}

