% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{snaiveForecast}
\alias{snaiveForecast}
\title{Seasonal naive forecast wrapper}
\usage{
snaiveForecast(x, h, level = 95, onlyfc = TRUE, ...)
}
\description{
Seasonal naive forecast wrapper
}

