% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{Start}
\alias{Start}
\title{Switcher funtion to choose frenquencies based on char dates}
\usage{
Start(freq, tdata)
}
\description{
Switcher funtion to choose frenquencies based on char dates
}

