% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{stsForecast}
\alias{stsForecast}
\title{Structural time series forecast wrapper}
\usage{
stsForecast(x, h, level = 95, onlyfc = TRUE, ...)
}
\description{
Structural time series forecast wrapper
}

