% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{summary.cvforecast}
\alias{summary.cvforecast}
\title{Default summary for Cross-validation forecasts}
\usage{
\method{summary}{cvforecast}(obj, digits = 4, plot = TRUE,
  accuracy.best = TRUE, cv.accuracy = FALSE)
}
\arguments{
\item{obj}{\code{cvforecast} object}

\item{digits}{number of digits for the default print output}

\item{plot}{if TRUE, plot forecast for best models}

\item{accuracy.best}{if TRUE, compute simple accuracy statistics for the best final models}

\item{cv.accuracy}{if TRUE, compute accuracy of the cross-validation points from 1 to horizon}
}
\description{
Summarizes \code{cvforecast} classed objects. The output is statistics about the best models fited by \code{cvforecast} function. It may include residual tests results and simple accuracy, cross-validation accuracy and a plot for the models.
}
\author{
LOPES, J. L.
}
\seealso{
\code{\link{cvforecast}}, \code{\link{accuracy}}, \code{\link{Mresid}}
}

