% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{switch.cvforecast}
\alias{switch.cvforecast}
\title{Switcher of time series forecast methods}
\usage{
switch.cvforecast(x, nmodelo, h, level = 95, onlyfc = FALSE)
}
\description{
Switcher of time series forecast methods
}

