% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{testObject}
\alias{testObject}
\title{Test if an object exists}
\usage{
testObject(object)
}
\arguments{
\item{object}{any object present into the environment}
}
\value{
TRE or FALSE
}
\description{
Test if an object exists
}

