% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{Try_error}
\alias{Try_error}
\title{Error handler improved}
\usage{
Try_error(code, silent = TRUE)
}
\arguments{
\item{code}{any stuff you want}

\item{silent}{if TRUE doesn't print results}
}
\description{
Improved verions of R error handler
}
\author{
LOPES, J. E.
}

