% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{yearsAgg}
\alias{yearsAgg}
\title{TSAgg Extra function 5}
\usage{
yearsAgg(data, process, multiple = NULL, na.rm = FALSE,
  from.first.obs = TRUE)
}
\description{
TSAgg Extra function 5
}

