% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{auto.arimaForecast}
\alias{auto.arimaForecast}
\title{auto.arima forecast wrapper}
\usage{
auto.arimaForecast(x, h, level = 95, onlyfc = TRUE, xreg = NULL,
  newxreg = NULL, ...)
}
\arguments{
\item{x}{'ts' data}

\item{h}{forecast horizon}

\item{level}{confidence level. Default is 0.95}

\item{onlyfc}{if TRUE return only forecasts, otherwise returns a full forecast classed object}

\item{xreg}{in ARIMA model with covariates it's must be provide}

\item{newxreg}{in ARIMA model with covariates it's must be provide for forecasting the covariates}

\item{...}{extra args, if needed.}
}
\value{
forecasts from ts data or an object of class 'forecast'
}
\description{
auto.arima forecast wrapper
}
\examples{
fit <- auto.arimaForecast(AirPassengers, h=10, onlyfc = FALSE)
plot(fit)
Mresid(fit)
tsSummary(fit)
}

