% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{correlation}
\alias{correlation}
\title{Correlations by pairs between two class of variables}
\usage{
correlation(tabela_y, tabela_x = NULL, method = "pearson", digits = 4,
  verbose = FALSE, corsimples = FALSE, ...)
}
\description{
Correlations by pairs between two class of variables
}

