% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvDeps.R, R/cvForecast.R
\name{cvforecast}
\alias{cvforecast}
\alias{package-cvforecast}
\title{Core function to compute multiple forecasts by Cross-Validation}
\usage{
cvforecast(tsdata, tsControl = cvForecastControl(), fcMethod = NULL, ...)
}
\arguments{
\item{tsdata}{data.frame type date-value, ts, mts or xts time series objects}

\item{tsControl}{generic contol with several args for the modelling process. See
\code{\link{cvForecastControl}}.}

\item{fcMethod}{accept the forecast method fefined by the user. This argument can be a string or a list, eg. fcMethod = "etsForecast" or a list as fcMethod = list("etsForecast", "HWsForecast"). If NULL, decision is made automatically.}

\item{...}{other arguments}
}
\value{
A list of class 'cvforecast' containing several objetcts from the forecasting process. It includes: betso models (less tahn 6), crossValidation statistics for all models, accuracy of all models, the control, etc. As below.
}
\description{
This function receives data and computes multiple forecasts by the technique of CrossValidation. The decision about the best model is based on tests as linearity, trend and fit accuracy.
}
\examples{
#Define cross validation parameters
myControl <- cvForecastControl(
 minObs = 14,
 stepSize = 5,
 maxHorizon = 30,
 fixedWindow=TRUE,
 preProcess=FALSE,
 ppMethod='guerrero',
 summaryFunc=tsSummary,
 cvMethod="MAPE",
 tsfrequency='day',
 OutlierClean=FALSE,
 dateformat='\%d/\%m/\%Y \%H:\%M:\%S')

#Paralell execution improves the processing time

#cl <- makeCluster(4, type='SOCK')
#registerDoParallel(cl)

#Load data

data(diario, package="cvforecast")
dadosd <- ConvertData(diario, dateformat='\%d/\%m/\%Y \%H:\%M:\%S', tsfrequency = "day", OutType="ts")
table(sapply(dadosd, class))
dim(dadosd)

#Looping example

FF <- llply(dadosd[,1:2], function(X) {
 fit <- try(cvforecast(X, myControl))
 if(class(fit) != "try-error") {
   #plot(fit)
   return(fit)
 } else NA
}, .progress = "time")

table(sapply(FF, class))
plot(FF[[1]])
sapply(FF, names)
#stopCluster(cl)
}
\keyword{crossValidation}
\keyword{series}
\keyword{time}

