% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{cvts2}
\alias{cvts2}
\title{Function to cross-validate a time series.}
\usage{
cvts2(x, FUN, tsControl = cvForecastControl(), progress = TRUE,
  packages = NULL, ...)
}
\description{
Function to cross-validate a time series.
}

