% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{cvts2}
\alias{cvts2}
\title{Function to cross-validate a time series}
\usage{
cvts2(x, FUN, tsControl = cvForecastControl(), progress = TRUE,
  packages = NULL, ...)
}
\arguments{
\item{x}{time series object of class 'ts'}

\item{FUN}{forecast wrapper function. These are some ones fc_sts, fc_hw, fc_tbats, fc_auto.arima, fc_ses, fc_mean, fc_holt, fc_bats, fc_ets, fc_lm, fc_theta, fc_rw, sfc_naive, fc_naive, fc_nnetar, fc_hws, fc_hwns and fc_hwes. This function works also in parallel very fast in multiple core computers.}

\item{tsControl}{Generic control for cross-validation process. See \code{\link{cvForecastControl}}.}

\item{progress}{if TRUE, show the progress bar.}

\item{packages}{extra R packages required by R. Default is NULL}

\item{...}{extra args, if needed}
}
\value{
list with information about then cross-validation like forecasts and accuracy
}
\description{
Main function to perform cross-validation. This function was firstly created by Zach Mayer (https://github.com/zachmayer/cv.ts) thanks, and adapted by LOPES. J. E. It
}
\examples{
# Control
tsControl <- cvForecastControl(
minObs = 14,
stepSize = 10,
maxHorizon = 30,
summaryFunc=tsSummary,
cvMethod="MAPE",
OutlierClean=FALSE)
#cl <- makeCluster(4, type='SOCK')
#registerDoParallel(cl)
x <- AirPassengers
fit <- cvts2(x, fc_auto.arima)
#stopCluster(cl)
}

