% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{descritiva}
\alias{descritiva}
\title{Descriptive analysis}
\usage{
descritiva(x, basic = TRUE, desc = TRUE, norm = FALSE, p = 0.95,
  dig = 6)
}
\arguments{
\item{x}{data object}

\item{basic}{if TRUE returns only basic descryptive statistics}

\item{desc}{if TRUE returns also hypothesis test for the mean}

\item{norm}{if TRUE returns also skewness and kurtosis}

\item{p}{level for confidence interval. Default is 0.95}

\item{digits}{output length of numeric data}

\item{...}{extra args, if needed.}
}
\description{
Perform Descriptive analysis on vector, matrix or data.frame
}
\examples{
descritiva(mtcars[,1:3])
}

