% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{forecastMethod}
\alias{forecastMethod}
\title{Choose best method based on trend and linearity}
\usage{
forecastMethod(x)

forecastMethod(x)
}
\description{
Choose best method based on trend and linearity

Find best forecast methods based on trend and linearity
}

