% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{lm_press_stat}
\alias{lm_press_stat}
\title{PRESS and other statistics from lm model}
\usage{
lm_press_stat(obj)
}
\arguments{
\item{obj}{A linear regression model (class 'lm'). Required.}
}
\description{
Returns the PRESS statistic (predictive residual sum of squares).
Useful for evaluating predictive power of regression models.
}
\examples{
fit <- step(lm(mpg~., data = mtcars), trace = 0)
lm_press_stat(fit)
}
\author{
LOPES, J. E.
}

