% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{nnetarForecast}
\alias{nnetarForecast}
\title{NNetar forecast wrapper}
\usage{
nnetarForecast(x, h, level = 95, onlyfc = TRUE, nn_p = 1, ...)
}
\arguments{
\item{x}{'ts' data}

\item{h}{forecast horizon}

\item{level}{confidence level. Default is 95%}

\item{onlyfc}{if TRUE return only forecasts, otherwise returns a full forecast classed object}

\item{...}{extra args, if needed.}
}
\value{
forecasts from ts data or an object of class 'forecast'
}
\description{
NNetar forecast wrapper
}
\examples{
fit <- nnetarForecast(AirPassengers, h=10, onlyfc = FALSE)
plot(fit)
Mresid(fit)
tsSummary(fit)
}

