% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{nparTrend}
\alias{nparTrend}
\title{Box and Cox tests and Ljung}
\usage{
nparTrend(x, ...)
}
\arguments{
\item{x}{'ts' data}
}
\value{
trend analysis statistics. See \code{\link{rkt}} for more information about tests
}
\description{
Use Mann-Kendall test (MK) and the Seasonal and the Regional Kendall Tests for trend (SKT and RKT) and Theil-Sen's slope estimator for checking trend
}
\examples{
nparTrend(AirPassengers)
}

