% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cvForecast.R
\name{plot.cvforecast}
\alias{plot.cvforecast}
\title{Default plot for Cross-validation forecast}
\usage{
plot.cvforecast(obj, ...)
}
\arguments{
\item{obj}{\code{cvforecast} object. See \code{\link{cvforecast}}}

\item{...}{extra args, if needed}
}
\description{
Plot \code{cvforecast} classed objects
}
\author{
LOPES, J. E.
}

