% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvForecastExtraFunctions.R
\name{Start}
\alias{Start}
\title{Compute starting values for ts transformation}
\usage{
Start(freq, tdata)
}
\arguments{
\item{freq}{data frequency for the dates. It works with  "year", "month", "day", "hour" and "min".}

\item{tdata}{data came from \code{TimeSeries} function}
}
\description{
This switcher is a internal function that chooses frequencies based on char dates. It's used to help transforming data into ts objects in R
}
\author{
LOPES, J. E.
}

